requirements.txt
# Core libraries
fastapi==0.114.1
uvicorn==0.30.6
gradio==4.44.0
apscheduler==3.10.4
pydantic==2.9.1
olefile==0.47
unstructured==0.15.12
unstructured[pdf]
python-magic-bin==0.4.14
pdfplumber==0.11.4

# Azure SDKs
azure-storage-blob==12.22.0
azure-storage-queue==12.11.0
azure-core==1.30.2
azure-search-documents==11.4.0
azure-identity==1.17.1

# LangChain & related tools
langchain==0.2.16
langchain-community==0.2.15
langchain-core==0.2.40
langchain-openai==0.1.24

# Other dependencies
tqdm==4.66.5
retrying==1.3.4

# For document processing
python-docx==1.1.2
PyPDF2==3.0.1

# For text-to-speech and speech-to-text
azure-cognitiveservices-speech==1.40.0

# For handling environments
python-dotenv==1.0.1